Extract portfolio transaction contracts
