+++ src/app/services/portfolio_service.py
@@ -37,11 +37,11 @@ from app.contracts.portfolio import (
     PortfolioReportingReadiness,
     PortfolioSummary,
     PortfolioTopPosition,
-    PortfolioTransactionLedgerResponse,
     PortfolioWorkflowResponse,
     PortfolioWorkspaceControlCapabilities,
     PortfolioWorkspaceResponse,
 )
+from app.contracts.portfolio_transactions import PortfolioTransactionLedgerResponse
 from app.precision_policy import (
     quantize_money,
     quantize_performance,
